parser/requirements.txt
beautifulsoup4==4.12.2
aiohttp==3.8.5
aiosqlite==0.19.0
python-dotenv==1.0.0
APScheduler==3.10.1
